tidy up usage of COnflictResolutionMechanism
