+++ src/main/scala/chatapp/ChatConflictResolutionMechanism.scala
@@ -1,24 +1,7 @@
 package chatapp
 
-import rover.rdo.conflict.{ConflictResolutionMechanism, ConflictedState, ResolvedMerge, SimpleConflictResolutionMechanism}
+import rover.rdo.conflict.provided.AppendIncomingChangesMergeResolve
 
-class ChatConflictResolutionMechanism(conflictedState: ConflictedState[List[ChatMessage]]) {
-
-//  override def resolveConflict(conflictedState: ConflictedState[List[ChatMessage]]): ResolvedMerge[List[ChatMessage]] = {
-//    val changesIncoming = conflictedState.changesIncomingRelativeToCommonAncestor
-//    var versionToApplyOn = conflictedState.serverVersion
-//
-//    for (operationToApply <- changesIncoming.asList){
-//      versionToApplyOn = operationToApply.appliedFunction(versionToApplyOn)
-//    }
-//
-//    val resultingVersion = versionToApplyOn
-//
-//    return new ResolvedMerge[List[ChatMessage]](conflictedState, resultingVersion.immutableState, this)
-//  }
-
-  def resolved: ResolvedMerge[List[ChatMessage]] = {
-    return new SimpleConflictResolutionMechanism[List[ChatMessage]].resolveConflict(conflictedState)
-  }
+class ChatConflictResolutionMechanism extends AppendIncomingChangesMergeResolve[List[ChatMessage]] {
 }
