+++ src/main/scala/votingapp/PollAppMergeConflictResolutionMechanism.scala
@@ -1,0 +1,7 @@
+package votingapp
+
+import rover.rdo.conflict.provided.AppendIncomingChangesMergeResolve
+
+class PollAppMergeConflictResolutionMechanism extends AppendIncomingChangesMergeResolve[Votes] {
+	
+}
